Extract encoding
