+++ lib/vrl/shared/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "vrl-shared"
+version = "0.1.0"
+authors = ["Vector Contributors <vector@timber.io>"]
+edition = "2018"
+publish = false
+license = "MPL-2.0"
+
+[dependencies]
+vrl-compiler = { path = "../compiler" }
+nom = { version = "6"}
+shared = { path = "../../shared", default-features = false, features = ["btreemap"]}
+
