+++ lib/vrl/shared/src/encode_logfmt.rs
@@ -1,0 +1,6 @@
+use crate::encode_key_value;
+use vrl_compiler::Value;
+
+pub fn encode<'a>(input: impl IntoIterator<Item = (String, Value)> + 'a) -> String {
+    encode_key_value::encode(input, &[], "=", " ", true)
+}
