+++ lib/vrl/shared/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod encode_key_value;
+pub mod encode_logfmt;
